+++ time4j-android/src/main/java/net/time4j/calendar/hindu/HinduMonth.java
@@ -240,7 +240,7 @@ public final class HinduMonth
      * @see     #getDisplayName(Locale)
      */
     public String getRasi(Locale locale) {
-        CalendarText names = CalendarText.getInstance("extra/hindu", locale);
+        CalendarText names = CalendarText.getInstance("hindu", locale);
         String rasi = names.getTextForms("R", IndianMonth.class).print(IndianMonth.valueOf(this.getRasi()));
 
         if (this.leap) { // should not happen because rasi is for solar calendars without leap months
@@ -428,7 +428,7 @@ public final class HinduMonth
     }
 
     private static String getAdhika(Locale locale) {
-        return CalendarText.getInstance("extra/hindu", locale).getTextForms().get("adhika") + " ";
+        return CalendarText.getInstance("hindu", locale).getTextForms().get("adhika") + " ";
     }
 
 }